+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## master
 
-* Playing asciicast from stdin: `cat demo.json | asciinema play -`
-* Playing asciicast from IPFS: `asciinema play ipfs://ipfs/QmcdXYJp6e4zNuimuGeWPwNMHQdxuqWmKx7NhZofQ1nw2V`
+* Added playback from stdin: `cat demo.json | asciinema play -`
+* Added playback from IPFS: `asciinema play ipfs://ipfs/QmcdXYJp6e4zNuimuGeWPwNMHQdxuqWmKx7NhZofQ1nw2V`
+* Added playback from asciicast page URL: `asciinema play https://asciinema.org/a/22124`
 * `-q/--quiet` option added to `rec` command
 * Fixed handling of partial UTF-8 sequences in recorded stdout
 * Final "exit" is now removed from recorded stdout